Move fake implementations of updater/downloader traits to own module
